+++ src/Models/CommonGeneral.php
@@ -16,4 +16,15 @@ class CommonGeneral extends NovaPoshta
 
         return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
     }
+
+    // не работает, требует JWT авторизацию
+    public function prolongateKey($ApiKey, $month = 12)
+    {
+        $this->calledMethod = 'prolongateApiKey';
+
+        $this->methodProperties['ApiKey'] = $ApiKey;
+        $this->methodProperties['prolongateMounthCount'] = $month;
+
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+    }
 }