+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 All notable changes to `NovaPoshta` will be documented in this file
 
 ### 2020-10-27
-- [Add][ND] `getCatalogCounterparty`
+- [Add][ND] `getCatalogCounterparty` in Counterparty
+- [Add] `getSettlements` in Address
+- [Add][ND] `prolongateKey` in CommonGeneral (NOT WORK)